+++ tests/test_estimation/test_calculate_weights.py
@@ -55,9 +55,7 @@ class TestCalcLowerE:
     def create_input_df(self):
         """Creates input df for test"""
 
-        input_cols = [
-            "employment",
-            "711"]
+        input_cols = ["employment", "711"]
         data = [
             [1, 10],
             [2, 5],
@@ -136,40 +134,8 @@ class TestCalcLowerSNoOutliers:
 # testing calculate_weighting_factors 709 to numeric with no missing vals
 # testing calculate_weighting_factors 709 to numeric with missing vals
 # testing calculate_weighting_factors with missing vals
-class TestCalcWeightMissingCol:
-    """Test for calculate_weighting_factors with missing outlier col"""
-
-    def create_input_df(self):
-        """Creates input df for test"""
-        input_cols = [
-            "reference",
-            "709"
-        ]
-
-        data = [
-            [1, "A"],
-            [2, "B"],
-            [2, "C"],
-            [4, "D"],
-            [1, "E"],
-        ]
-
-        input_df = pd.DataFrame(data=data, columns=input_cols)
-        return input_df
-
-    def test_calculate_weighting_factors_missing_col(self):
-        """Test for calculate_weighting_factors with missing outlier col"""
-        input_df = self.create_input_df()
-
-        with pytest.raises(
-            ValueError, match=r"The column essential 'outlier' is missing"
-        ):
-            calw.calculate_weighting_factors(input_df)
-
-
-class TestCalcWeightFactor:
-    """Test for calculate_weighting_factors for filter
-    and np.nan taken out of calculation"""
+class TestCalcWeightFactors:
+    """Tests for calculate_weighting_factors function."""
 
     def create_input_df(self):
         """Creates input df for test"""
@@ -177,9 +143,6 @@ class TestCalcWeightFactor:
             "reference",
             "instance",
             "cellnumber",
-            "selectiontype",
-            "status",
-            "formtype",
             "709",
             "uni_count",
             "uni_employment",
@@ -188,19 +151,17 @@ class TestCalcWeightFactor:
         ]
 
         data = [
-            [1, 0, 1, "P", "Clear", "0006", "12", 20, 5000, 66, True],
-            [2, 0, 2, "P", "Clear - overridden", "0006", 14, 4, 5000, 77, False],
-            [2, 1, 2, "P", "Clear", "0006", 16, 4, 5000, 77, False],
-            [3, 0, 1, "P", "999", "0006", 1, 20, 5000, 11, False],
-            [4, 0, 4, "P", "Clear", "0006", 18, 3, 5000, 88, False],
-            [5, 0, 2, "P", "Clear - overridden", "0001", 14, 4, 5000, 22, False],
-            [6, 0, 1, "P", "Clear", "0006", 10, 20, 5000, 7, False],
-            [7, 0, 5, "P", "Clear", "0006", 20, 50, 5000, 20, False],
-            [7, 1, 5, "P", "Clear", "0006", 20, 50, 5000, 20, False],
-            [8, 1, 2, "P", "Clear", "0006", np.nan, 4, 5000, 7, False],
-            [9, 0, 1, "P", "Clear", "0006", 5, 20, 5000, 44, False],
-            [10, 0, 1, "P", "Clear", "0006", 10, 20, 5000, 44, False],
-            [11, 0, 5, "X", "Clear", "0006", 20, 50, 5000, 20, False],
+            [1, 0, 1, "12", 20, 5000, 66, True],
+            [2, 0, 2, 14, 4, 5000, 77, False],
+            [3, 0, 1, 1, 20, 5000, 11, False],
+            [4, 0, 4, 18, 3, 5000, 88, False],
+            [5, 0, 2, 14, 4, 5000, 22, False],
+            [6, 0, 1, 10, 20, 5000, 7, False],
+            [7, 0, 5, 20, 50, 5000, 20, False],
+            [8, 0, 2, np.nan, 4, 5000, 7, False],
+            [9, 0, 1, 5, 20, 5000, 44, False],
+            [10, 0, 1, 10, 20, 5000, 44, False],
+            [11, 0, 5, 20, 50, 5000, 20, False],
         ]
 
         input_df = pd.DataFrame(data=data, columns=input_cols)
@@ -212,9 +173,6 @@ class TestCalcWeightFactor:
             "reference",
             "instance",
             "cellnumber",
-            "selectiontype",
-            "status",
-            "formtype",
             "709",
             "uni_count",
             "uni_employment",
@@ -225,19 +183,17 @@ class TestCalcWeightFactor:
         ]
 
         data = [
-            [1, 0, 1, "P", "Clear", "0006", "12", 20, 5000, 66, True, 1.0, 1.0],
-            [2, 0, 2, "P", "Clear - overridden", "0006", 14, 4, 5000, 77, False, 4.0, 16.2],
-            [2, 1, 2, "P", "Clear", "0006", 16, 4, 5000, 77, False, 4.0, 16.2],
-            [3, 0, 1, "P", "999", "0006", 1, 20, 5000, 11, False, 6.3, 8.2],
-            [4, 0, 4, "P", "Clear", "0006", 18, 3, 5000, 88, False, 3.0, 18.9],
-            [5, 0, 2, "P", "Clear - overridden", "0001", 14, 4, 5000, 22, False, 1.0, 1.0],
-            [6, 0, 1, "P", "Clear", "0006", 10, 20, 5000, 7, False, 6.3, 8.2],
-            [7, 0, 5, "P", "Clear", "0006", 20, 50, 5000, 20, False, 50.0, 5.0],
-            [7, 1, 5, "P", "Clear", "0006", 20, 50, 5000, 20, False, 50.0, 5.0],
-            [8, 1, 2, "P", "Clear", "0006", np.nan, 4, 5000, 7, False, 4.0, 16.2],
-            [9, 0, 1, "P", "Clear", "0006", 5, 20, 5000, 44, False, 6.3, 8.2],
-            [10, 0, 1, "P", "Clear", "0006", 10, 20, 5000, 44, False, 6.3, 8.2],
-            [11, 0, 5, "X", "Clear", "0006", 20, 50, 5000, 20, False, 1.0, 1.0],
+            [1, 0, 1, "12", 20, 5000, 66, True, 1.0, 1.0],
+            [2, 0, 2, 14, 4, 5000, 77, False, 1.3, 35.4],
+            [3, 0, 1, 1, 20, 5000, 11, False, 4.8, 9.8],
+            [4, 0, 4, 18, 3, 5000, 88, False, 3.0, 18.9],
+            [5, 0, 2, 14, 4, 5000, 22, False, 1.3, 35.4],
+            [6, 0, 1, 10, 20, 5000, 7, False, 4.8, 9.8],
+            [7, 0, 5, 20, 50, 5000, 20, False, 25.0, 5.0],
+            [8, 0, 2, np.nan, 4, 5000, 7, False, 1.3, 35.4],
+            [9, 0, 1, 5, 20, 5000, 44, False, 4.8, 9.8],
+            [10, 0, 1, 10, 20, 5000, 44, False, 4.8, 9.8],
+            [11, 0, 5, 20, 50, 5000, 20, False, 25.0, 5.0],
         ]
         expected_df = pd.DataFrame(data=data, columns=expected_cols)
         return expected_df
@@ -257,18 +213,17 @@ class TestCalcWeightFactor:
         ]
 
         data = [
-            [1.0, 20.0, 4.0, 1.0, 5000, 161, 66, 6.3, 8.2],
-            [2.0, 4.0, 1.0, 0.0, 5000, 77, 0, 4.0, 16.23],
-            [4.0, 3.0, 1.0, 0.0, 5000, 88, 0, 3.0, 18.93],
-            [5.0, 50.0, 1.0, 0.0, 5000, 20, 0, 50.0, 5.0],
+            [1, 20, 5, 1, 5000, 172, 66, 4.8, 9.8],
+            [2, 4, 3, 0, 5000, 106, 0, 1.3, 35.4],
+            [4, 3, 1, 0, 5000, 88, 0, 3.0, 18.9],
+            [5, 50, 2, 0, 5000, 40, 0, 25.0, 5.0],
         ]
 
         expected_qa_df = pd.DataFrame(data=data, columns=expected_qa_cols)
         return expected_qa_df
 
-    def test_calculate_weighting_factors_filter(self):
-        """Test for calculate_weighting_factors for filter
-        and np.nan taken out of calculation"""
+    def test_calculate_weighting_factors_g_weight_true(self):
+        """Test for calculate_weighting_factors with g_weight set to True"""
 
         input_df = self.create_input_df()
         expected_df = self.create_expected_output()
@@ -276,13 +231,14 @@ class TestCalcWeightFactor:
         print(expected_qa_df.columns)
         result_df, result_qa_df = calw.calculate_weighting_factors(input_df)
 
-        result_qa_df["a_weight"] = result_qa_df["a_weight"].round(1)
-        result_qa_df["g_weight"] = result_qa_df["g_weight"].round(1)
+        for df in [result_qa_df, result_df]:
+            df["a_weight"] = df["a_weight"].round(1)
+            df["g_weight"] = df["g_weight"].round(1)
 
         assert_frame_equal(result_df, expected_df, check_exact=False, rtol=0.01)
         assert_frame_equal(result_qa_df, expected_qa_df, check_exact=False, rtol=0.01)
 
-    def test_calculate_weighting_factors_filter(self):
+    def test_calculate_weighting_factors_g_weight_false(self):
         """Test for calculate_weighting_factors for filter
         and np.nan taken out of calculation"""
 
@@ -290,16 +246,21 @@ class TestCalcWeightFactor:
         expected_df = self.create_expected_output()
         expected_qa_df = self.create_expected_qa()
 
-        result_df, result_qa_df = calw.calculate_weighting_factors(input_df)
+        expected_df = expected_df.drop(columns=["g_weight"])
+        expected_qa_df = expected_qa_df.drop(
+            columns=[
+                "E - uni_employment",
+                "e - sum of employment in cell",
+                "s - sum of employment outliers in cell",
+                "g_weight"
+            ]
+        )
 
-        # list of DataFrames and columns to round
-        dfs = [result_qa_df, result_df, expected_df, expected_qa_df]
-        columns = ["a_weight", "g_weight"]
+        result_df, result_qa_df = calw.calculate_weighting_factors(input_df, calc_g_weights=False)
 
         # Round specified columns in each DataFrame
-        for df in dfs:
-            for col in columns:
-                df[col] = df[col].round(1)
+        for df in [result_qa_df, result_df]:
+            df["a_weight"] = df["a_weight"].round(1)
 
         # Ensure both DataFrames have the same data type for the "709" column
         result_df["709"] = result_df["709"].astype(float)
@@ -308,144 +269,12 @@ class TestCalcWeightFactor:
         assert_frame_equal(result_df, expected_df, check_exact=False, rtol=0.01, check_dtype=False)
         assert_frame_equal(result_qa_df, expected_qa_df, check_exact=False, rtol=0.01, check_dtype=False)
 
-
-class TestCalcWeightsSecondCheck:
-    """Test for calculate_weighting_factors for filter
-    and np.nan taken out of calculation"""
-
-    def create_input_df(self):
-        """Create an input dataframe for the test."""
-        input_columns = [
-            "reference",
-            "instance",
-            "cellnumber",
-            "formtype",
-            "709",
-            "selectiontype",
-            "status",
-            "uni_count",
-            "uni_employment",
-            "employment",
-            "outlier",
-        ]
-
-        data = [
-            [1, 0, 1, "0006", 12.0, "P", "Clear", 20, 2000, 7, True],
-            [1, 1, 1, "0006", 20.0, "P", "Clear", 20, 2000, 7, True],
-            [2, 0, 2, "0006", 16.0, "P", "Clear - overridden", 4, 2000, 77, False],
-            [2, 1, 2, "0006", 16.0, "P", "Clear - overridden", 4, 2000, 77, False],
-            [3, 0, 1, "0006", np.nan, "P", "Clear - overridden", 20, 2000, 11, False],
-            [4, 0, 4, "0006", 18.0, "P", "Clear", 30, 2000, 88, False],
-            [5, 1, 3, "0001", 14.0, "C", "Clear - overridden", 50, 2000, 22, False],
-            [6, 0, 1, "0006", 10.0, "P", "Clear", 20, 2000, 50, False],
-        ]
-
-        input_df = pd.DataFrame(data=data, columns=input_columns)
-        return input_df
-
-
-    def create_expected_df(self):
-        """Create an expected dataframe for the test."""
-        expected_columns = [
-            "reference",
-            "instance",
-            "cellnumber",
-            "formtype",
-            "709",
-            "selectiontype",
-            "status",
-            "uni_count",
-            "uni_employment",
-            "employment",
-            "outlier",
-            "a_weight",
-            "g_weight",
-        ]
-
-        data = [
-            [1, 0, 1, "0006", 12.0, "P", "Clear", 20, 2000, 7, True, 1, 1.0],
-            [1, 1, 1, "0006", 20.0, "P", "Clear", 20, 2000, 7, True, 1, 1.0],
-            [2, 0, 2, "0006", 16.0, "P", "Clear - overridden", 4, 2000, 77, False, 4, 6.494],
-            [2, 1, 2, "0006", 16.0, "P", "Clear - overridden", 4, 2000, 77, False, 4, 6.494],
-            [3, 0, 1, "0006", np.nan, "P", "Clear - overridden", 20, 2000, 11, False, 19, 2.098],
-            [4, 0, 4, "0006", 18.0, "P", "Clear", 30, 2000, 88, False, 30, 0.758],
-            [5, 1, 3, "0001", 14.0, "C", "Clear - overridden", 50, 2000, 22, False, 1, 1.0],
-            [6, 0, 1, "0006", 10.0, "P", "Clear", 20, 2000, 50, False, 19, 2.098],
-        ]
-
-        expected_df = pd.DataFrame(data=data, columns=expected_columns)
-        return expected_df
-
-
-    def create_expected_qa(self):
-        """Creates expected qa df for test"""
-        expected_qa_cols = [
-            "Cell Number",
-            "N - uni_count",
-            "n - num clear records in cell",
-            "o - num outliers in cell",
-            "E - uni_employment",
-            "e - sum of employment in cell",
-            "s - sum of employment outliers in cell",
-            "a_weight",
-            "g_weight"
-        ]
-
-        data = [
-            [1, 20, 2, 1.0, 2000, 57, 7, 19, 2.098],
-            [2, 4, 1, 0.0, 2000, 77, 0, 4, 6.494],
-            [4, 30, 1, 0.0, 2000, 88, 0, 30, 0.758],
-        ]
-
-        expected_qa_df = pd.DataFrame(data=data, columns=expected_qa_cols)
-        return expected_qa_df
-
-    def test_calculate_weighting_factors_with_missing_vals(self):
-        """Test for calculate_weighting_factors for filter
-        and np.nan taken out of calculation"""
-
-        input_df = self.create_input_df()
-        expected_df = self.create_expected_df()
-        expected_qa_df = self.create_expected_qa()
-
-        result_df, result_qa_df = calw.calculate_weighting_factors(input_df)
-
-        # list of DataFrames and columns to round
-        dfs = [result_qa_df, result_df, expected_df, expected_qa_df]
-        columns = ["a_weight", "g_weight"]
-
-        # Round specified columns in each DataFrame
-        for df in dfs:
-            for col in columns:
-                df[col] = df[col].round(3)
-
-
-        assert_frame_equal(
-            result_df, expected_df, check_exact=False, rtol=0.01, check_dtype=False
-        )
-        assert_frame_equal(
-            result_qa_df,
-            expected_qa_df,
-            check_exact=False,
-            rtol=0.01,
-            check_dtype=False,
-        )
-
-
-# One tests for outlier_weights:
-# test that all appropriate rows are given an a_weight = 1.0
-
-
 class TestOutlierWeight:
     """Test for outlier_weights."""
 
     def create_input_df(self):
         """Creates input df for test"""
-        input_cols = [
-            "reference",
-            "outlier",
-        ]
-
+        input_cols = ["reference","outlier"]
         data = [
             [1, True],
             [2, False],